Move internal target to meta
Internal target is part of the metas, as it is build
from the documents, and overlapping for the full set.
